requirements.txt
# Core dependencies
python-dotenv>=1.0.0
slack-sdk>=3.21.3
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
trello==0.9.7
PyYAML>=6.0.1

# Email processing
python-imap==1.0.0
email-validator==2.0.0.post2
python-magic==0.4.27

# Web research
selenium==4.11.2
webdriver-manager==4.0.0
requests==2.31.0
html2text==2020.1.16
readability-lxml==0.8.1

# Data processing
pandas==2.0.3
numpy>=1.24.3
nltk==3.8.1
spacy==3.6.0
python-dateutil>=2.8.2

# Document management
python-docx==0.8.11
PyPDF2==3.0.1
openpyxl==3.1.2

# Utilities
tqdm==4.65.0
colorlog==6.7.0
tenacity==8.2.2

# Voice Input Requirements
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
pynput>=1.7.6
vosk>=0.3.45
sounddevice>=0.4.6

# File Transfer Requirements
markdown>=3.3.4

# System Interaction Requirements
psutil>=5.9.0  # For process management and system monitoring 

# Browser Automation Requirements

# Core Request Requirements
backoff>=2.2.1
cachetools>=5.3.0
ratelimit>=2.2.1 

# Added from the code block
openai==1.3.7 

# Added from the code block
scikit-learn>=1.3.0 